Move camera control functions to CameraControl
