+++ src/FaceDetection.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Utils\CornerBoundingBox.cs" />
     <Compile Include="Utils\DelegateCommand.cs" />
     <Compile Include="Utils\FaceBoundingBox.cs" />
+    <Compile Include="Utils\CameraControl.cs" />
     <Compile Include="Utils\ImageUtils.cs" />
     <Compile Include="ViewModels\MainPageViewModel.cs" />
   </ItemGroup>
