+++ src/ViewModels/MainPageViewModel.cs
@@ -24,28 +24,24 @@ namespace FaceDetection.ViewModels
 {
     public class MainPageViewModel : BaseNotifyPropertyChanged
     {
-        private FrameModel _frameModel = new FrameModel();
-
-        private bool _isInitialized;
-        public bool IsPreviewing { get; set; }
-        public MediaCapture MediaCapture { get; set; }
-
+        public FrameModel FrameModel { get; } = new FrameModel();
+        public CameraControl CameraControl { get; } = new CameraControl();
         private int _processingFlag;
-        private MediaFrameReader _frameReader;
-
-        // TODO: Create Config to load modelFileName
-        private string _modelFileName = "version-RFB-320.onnx";
 
         private IFaceDetector _faceDetector;
         private bool _isFaceDetectionEnabled = false;
-        public bool IsFaceDetectionEnabled {
+        // TODO: Create Config to load modelFileName
+        private string _modelFileName = "version-RFB-320.onnx";
+        
+        public event FaceDetectedEventHandler FaceDetected;
+        public bool IsFaceDetectionEnabled
+        {
             get => _isFaceDetectionEnabled;
             set
             {
                 SetProperty(ref _isFaceDetectionEnabled, value);
             }
         }
-        public event FaceDetectedEventHandler FaceDetected;
 
         private FocalLengthBasedDistanceEstimator _distanceEstimator
             = new FocalLengthBasedDistanceEstimator();
@@ -54,28 +50,6 @@ namespace FaceDetection.ViewModels
         {
             SubscribeEvents();
             Task.Run(LoadModelAsync);
-        }
-
-        private void SubscribeEvents()
-        {
-            _frameModel.PropertyChanged += _frameModel_PropertyChanged;
-            PropertyChanged += MainPageViewModel_PropertyChanged;
-        }
-
-        private void MainPageViewModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(IsFaceDetectionEnabled))
-            {
-                PerformFaceDetection();
-            }
-        }
-
-        private void _frameModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(_frameModel.SoftwareBitmap))
-            {
-                PerformFaceDetection();
-            }
         }
 
         public void OnFrameArrived(MediaFrameReader sender, MediaFrameArrivedEventArgs args)
@@ -87,83 +61,11 @@ namespace FaceDetection.ViewModels
                 {
                     if (bmp != null)
                     {
-                        _frameModel.SoftwareBitmap = bmp;
+                        FrameModel.SoftwareBitmap = bmp;
                     }
                 }
                 Interlocked.Exchange(ref _processingFlag, 0);
             }
-        }
-
-        public async Task InitCameraAsync()
-        {
-            if (!_isInitialized)
-            {
-                await InitMediaCaptureAsync();
-            }
-            if (!_isInitialized) return;
-
-            await InitFrameReaderAsync();
-        }
-
-        public async Task StartPreviewAsync()
-        {
-            try
-            {
-                await MediaCapture.StartPreviewAsync();
-                IsPreviewing = true;
-            }
-            catch (FileLoadException)
-            {
-                Debug.WriteLine("Another app has exclusive access");
-            }
-        }
-
-        public async Task StopPreviewAsync()
-        {
-            IsPreviewing = false;
-            await MediaCapture.StopPreviewAsync();
-            await _frameReader.StopAsync();
-        }
-
-        private async Task InitMediaCaptureAsync()
-        {
-            var cameraDevice = await FindCameraDeviceByPanelAsync(Windows.Devices.Enumeration.Panel.Front);
-            if (cameraDevice == null)
-            {
-                Debug.WriteLine("No camera device found");
-                return;
-            }
-
-            // Set Memory Preference to CPU to guarantee SoftwareBitmap property is non-null
-            // Otherwise use Direct3DSurface property
-            // https://docs.microsoft.com/en-us/uwp/api/windows.media.capture.frames.videomediaframe.softwarebitmap?view=winrt-19041
-            MediaCapture = new MediaCapture();
-            var settings = new MediaCaptureInitializationSettings
-            {
-                VideoDeviceId = cameraDevice.Id, MemoryPreference = MediaCaptureMemoryPreference.Cpu
-            };
-
-            try
-            {
-                await MediaCapture.InitializeAsync(settings);
-                _isInitialized = true;
-            }
-            catch (UnauthorizedAccessException)
-            {
-                Debug.WriteLine("The app was denied access to the camera");
-            }
-        }
-
-        private async Task InitFrameReaderAsync()
-        {
-            var frameSource = MediaCapture.FrameSources.Where(
-                source => source.Value.Info.SourceKind == MediaFrameSourceKind.Color)
-                .First();
-            _frameReader = await MediaCapture.CreateFrameReaderAsync(frameSource.Value, MediaEncodingSubtypes.Rgb32);
-
-            // Setup handler for frames
-            _frameReader.FrameArrived += OnFrameArrived;
-            await _frameReader.StartAsync();
         }
 
         private async void PerformFaceDetection()
@@ -171,25 +73,12 @@ namespace FaceDetection.ViewModels
             if (!_isFaceDetectionEnabled || _faceDetector == null) return;
             else if (_faceDetector != null && !_faceDetector.IsModelLoaded()) return;
 
-            SoftwareBitmap bmp = _frameModel.SoftwareBitmap;
+            SoftwareBitmap bmp = FrameModel.SoftwareBitmap;
             if (bmp == null) return;
             Mat img = ImageUtils.ConvertSoftwareBitmapToMat(bmp);
             if (img == null) return;
             await _faceDetector.Detect(img);
             img.Dispose();
-        }
-
-        private async Task LoadModelAsync()
-        {
-            StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri($"ms-appx:///Assets/{_modelFileName}"));
-            _faceDetector = new UltraFaceDetector();
-            _faceDetector.LoadModel(file);
-            _faceDetector.FaceDetected += _faceDetector_FaceDetected;
-        }
-
-        private void _faceDetector_FaceDetected(object sender, IReadOnlyList<FaceBoundingBox> faceBoundingBoxes, System.Drawing.Size originalSize)
-        {
-            FaceDetected?.Invoke(sender, faceBoundingBoxes, originalSize);
         }
 
         public FaceBoundingBox ScaleBoundingBox(FaceBoundingBox origBB, System.Drawing.Size originalSize)
@@ -210,18 +99,49 @@ namespace FaceDetection.ViewModels
             return _distanceEstimator.ComputeDistance(bb);
         }
 
-        public async void CacheImageFromStreamAsync(IRandomAccessStream fileStream)
+        private async Task LoadModelAsync()
         {
-            BitmapDecoder decoder = await BitmapDecoder.CreateAsync(fileStream);
-            SoftwareBitmap softwareBitmap = await decoder.GetSoftwareBitmapAsync(BitmapPixelFormat.Bgra8, BitmapAlphaMode.Ignore);
-            _frameModel.SoftwareBitmap = softwareBitmap;
+            StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri($"ms-appx:///Assets/{_modelFileName}"));
+            _faceDetector = new UltraFaceDetector();
+            _faceDetector.LoadModel(file);
+            _faceDetector.FaceDetected += _faceDetector_FaceDetected;
         }
 
-        private static async Task<DeviceInformation> FindCameraDeviceByPanelAsync(Windows.Devices.Enumeration.Panel desiredPanel)
+        private void SubscribeEvents()
         {
-            var allVideoDevices = await DeviceInformation.FindAllAsync(DeviceClass.VideoCapture);
-            DeviceInformation desiredDevice = allVideoDevices.FirstOrDefault(x => x.EnclosureLocation != null && x.EnclosureLocation.Panel == desiredPanel);
-            return (desiredDevice == null) ? allVideoDevices.FirstOrDefault() : desiredDevice;
+            FrameModel.PropertyChanged += _frameModel_PropertyChanged;
+            PropertyChanged += MainPageViewModel_PropertyChanged;
         }
+
+        private void MainPageViewModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            if (e.PropertyName == nameof(IsFaceDetectionEnabled))
+            {
+                PerformFaceDetection();
+            }
+        }
+
+        private void _frameModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            if (e.PropertyName == nameof(FrameModel.SoftwareBitmap))
+            {
+                PerformFaceDetection();
+            }
+        }
+
+        private void _faceDetector_FaceDetected(object sender, IReadOnlyList<FaceBoundingBox> faceBoundingBoxes, System.Drawing.Size originalSize)
+        {
+            FaceDetected?.Invoke(sender, faceBoundingBoxes, originalSize);
+        }
+
+        public async Task InitCameraAsync()
+        {
+            await CameraControl.InitCameraAsync();
+            CameraControl.RegisterFrameArrivedCallback(OnFrameArrived);
+        }
+
+        public async Task StartPreviewAsync() => await CameraControl.StartPreviewAsync();
+
+        public async Task StopPreviewAsync() => await CameraControl.StopPreviewAsync();
     }
 }
